refactor: account security page individual tab components
